refactor: sidebar 컴포넌트 추출
